move image specific code to utils for easier refactoring
